+++ stream-chat-android-ui-common/src/test/kotlin/com/getstream/sdk/chat/messages/MessageListControllerTests.kt
@@ -357,7 +357,7 @@ internal class MessageListControllerTests {
         }
 
         fun get(
-            dateSeparatorHandler: DateSeparatorHandler = DateSeparatorHandler.getDefaultDateSeparator(),
+            dateSeparatorHandler: DateSeparatorHandler = DateSeparatorHandler.getDefaultDateSeparatorHandler(),
             deletedMessageVisibility: DeletedMessageVisibility = DeletedMessageVisibility.ALWAYS_VISIBLE
         ): MessageListController {
             return MessageListController(